+++ web/client/src/library/components/documentation/Documentation.tsx
@@ -1,19 +1,18 @@
-import { type Model } from '@api/client'
+import React from 'react'
 import CodeEditor, {
   useSQLMeshModelExtensions,
 } from '@components/editor/EditorCode'
-import { useStoreEditor } from '@context/editor'
-import { useStoreLineage } from '@context/lineage'
 import { Disclosure, Tab } from '@headlessui/react'
 import { MinusCircleIcon, PlusCircleIcon } from '@heroicons/react/24/solid'
 import { EnumFileExtensions } from '@models/file'
 import { isFalse, isString, isTrue, toDateFormat } from '@utils/index'
 import clsx from 'clsx'
-import React, { useEffect } from 'react'
 import { useNavigate } from 'react-router-dom'
 import { EnumRoutes } from '~/routes'
+import { ModelSQLMeshModel } from '@models/sqlmesh-model'
+import { ModelColumn } from '@components/graph/Graph'
 
-const Documentation = function Documentation({
+const Documantation = function Documantation({
   model,
   withModel = true,
   withDescription = true,
@@ -21,7 +20,7 @@ const Documentation = function Documentation({
   withCode = true,
   withQuery = true,
 }: {
-  model: Model
+  model: ModelSQLMeshModel
   withCode?: boolean
   withQuery?: boolean
   withModel?: boolean
@@ -29,25 +28,12 @@ const Documentation = function Documentation({
   withColumns?: boolean
 }): JSX.Element {
   const navigate = useNavigate()
-  const lineage = useStoreEditor(s => s.previewLineage)
 
-  const setColumns = useStoreLineage(s => s.setColumns)
-  const clearActiveEdges = useStoreLineage(s => s.clearActiveEdges)
-
-  const modelExtensions = useSQLMeshModelExtensions(
-    model.path,
-    lineage,
-    model => {
-      navigate(`${EnumRoutes.IdeDocsModels}?model=${model.name}`, {
-        state: { model },
-      })
-    },
-  )
-
-  useEffect(() => {
-    setColumns(undefined)
-    clearActiveEdges()
-  }, [model])
+  const modelExtensions = useSQLMeshModelExtensions(model.path, model => {
+    navigate(
+      `${EnumRoutes.IdeDocsModels}/${ModelSQLMeshModel.encodeName(model.name)}`,
+    )
+  })
 
   return (
     <Container>
@@ -68,6 +54,10 @@ const Documentation = function Documentation({
             <DetailsItem
               name="Dialect"
               value={model.dialect}
+            />
+            <DetailsItem
+              name="Type"
+              value={model.type}
             />
             {Object.entries(model.details ?? {}).map(([key, value]) => (
               <DetailsItem
@@ -90,16 +80,25 @@ const Documentation = function Documentation({
       )}
       {withColumns && (
         <Section headline="Columns">
-          <ul className="px-2 w-full">
+          <ul className="w-full">
             {model.columns.map(column => (
-              <DetailsItem
+              <ModelColumn
                 key={column.name}
-                name={column.name}
-                value={column.type}
-                isHighlighted={true}
+                model={model}
+                column={column}
+                disabled={model.type === 'python'}
+                className="px-2 py-1 rounded-md border-b border-primary-10 mb-1"
               >
-                {column.description}
-              </DetailsItem>
+                {({ column, disabled }) => (
+                  <ModelColumn.Display
+                    columnName={column.name}
+                    columnType={column.type}
+                    columnDescription={column.description}
+                    isHighlighted={true}
+                    disabled={disabled}
+                  />
+                )}
+              </ModelColumn>
             ))}
           </ul>
         </Section>
@@ -144,7 +143,7 @@ const Documentation = function Documentation({
                       Compiled Query
                     </Tab>
                   </Tab.List>
-                  <Tab.Panels className="h-full w-full overflow-hidden p-2 bg-neutral-10 mt-4 rounded-lg">
+                  <Tab.Panels className="h-full w-full overflow-hidden bg-neutral-10 mt-4 rounded-lg">
                     <Tab.Panel
                       className={clsx(
                         'flex flex-col w-full h-full pt-4 relative px-2 overflow-hidden',
@@ -258,7 +257,7 @@ function Section({
                 )}
               </div>
             </Disclosure.Button>
-            <Disclosure.Panel className="px-4 pb-2 text-sm">
+            <Disclosure.Panel className="pb-2 text-sm overflow-hidden">
               <div className="px-2">{children}</div>
             </Disclosure.Panel>
           </>
@@ -315,7 +314,7 @@ function DetailsItem({
   )
 }
 
-Documentation.NotFound = NotFound
-Documentation.Container = Container
+Documantation.NotFound = NotFound
+Documantation.Container = Container
 
-export default Documentation
+export default Documantation
